+++ src/console/controllers/MakeController.php
@@ -8,13 +8,13 @@
  * @copyright Copyright (c) 2020 Percipio Global Ltd.
  */
 
-namespace percipioglobal\notifications\console\controllers;
-
-use craft\helpers\FileHelper;
-use League\Flysystem\Adapter\Local;
-use League\Flysystem\Filesystem;
+namespace percipiolondon\notifications\console\controllers;
 
 use Craft;
+
+use craft\helpers\FileHelper;
+use yii\base\ErrorException;
+use yii\base\Exception;
 use yii\base\Module;
 use yii\console\Controller;
 
@@ -52,7 +52,6 @@ class MakeController extends Controller
         parent::__construct($id, $module, $config);
     }
 
-
     /**
      * Handle notifications/default console commands
      *
@@ -61,11 +60,11 @@ class MakeController extends Controller
      *
      * @param string $name
      *
-     * @return mixed
-     * @throws \yii\base\Exception
-     * @throws \yii\base\ErrorException
+     * @return bool
+     * @throws Exception
+     * @throws ErrorException
      */
-    public function actionIndex($name)
+    public function actionIndex(string $name): bool
     {
         // First we will check to see if the class already exists. If it does, we don't want
         // to create the class and overwrite the user's code. So, we will bail out so the
@@ -77,33 +76,36 @@ class MakeController extends Controller
 
         // Make sure the directory exists
         $dir = CRAFT_BASE_PATH . '/notifications';
-        if (! is_dir($dir)) {
+        if (!is_dir($dir)) {
             FileHelper::createDirectory($dir);
         }
 
         FileHelper::writeToFile($dir . "/{$name}.php", $this->buildClass($name));
 
         $this->stdout("Notification created successfully.");
+
+        return true;
     }
 
     /**
      * Get the stub file for the generator.
      *
      * @return string
-     * @throws \yii\base\Exception
+     * @throws Exception
      */
-    protected function getStub()
+    protected function getStub(): string
     {
-        return Craft::$app->path->getVendorPath() . '/percipioglobal/craft-notifications/src/notification.stub';
+        // return Craft::$app->path->getVendorPath() . '/percipioglobal/craft-notifications/src/notification.stub';
+        return dirname(dirname(__DIR__)) . '/notification.stub';
     }
 
     /**
      * Determine if the class already exists.
      *
-     * @param  string  $rawName
+     * @param string $rawName
      * @return bool
      */
-    protected function alreadyExists($rawName)
+    protected function alreadyExists(string $rawName): bool
     {
         return file_exists(CRAFT_BASE_PATH . "/notifications/{$rawName}.php");
     }
@@ -114,7 +116,7 @@ class MakeController extends Controller
      * @param  string $name
      *
      * @return string
-     * @throws \yii\base\Exception
+     * @throws Exception
      */
     protected function buildClass($name)
     {