+++ src/models/Notification.php
@@ -8,14 +8,11 @@
  * @copyright Copyright (c) 2020 Percipio Global Ltd.
  */
 
-namespace percipioglobal\notifications\models;
-
-use craft\base\ElementInterface;
-use craft\base\Model;
-use percipioglobal\notifications\Notifications;
+namespace percipiolondon\notifications\models;
 
 use Craft;
-use craft\db\ActiveRecord;
+use craft\base\Model;
+
 use yii\base\Event;
 
 /**
@@ -44,12 +41,7 @@ class Notification extends Model
      */
     public function __construct($config)
     {
-        if ($config instanceof Event) {
-            Craft::$app->getDeprecator()->log('Notification::__construct()', 'Passing a yii\base\Event to Notification::__construct() has been deprecated. Pass a config array with a “event” value instead.');
-
-            $config = ['event' => $config];
-        }
-
+        $config = ['event' => $config];
         parent::__construct($config);
     }
 
@@ -66,7 +58,7 @@ class Notification extends Model
      *
      * @return array
      */
-    public function via()
+    public function via(): array
     {
         return [];
     }