+++ src/services/NotificationsService.php
@@ -8,21 +8,25 @@
  * @copyright Copyright (c) 2018 Rias
  */
 
-namespace percipioglobal\notifications\services;
+namespace percipiolondon\notifications\services;
 
-use craft\elements\User;
-use craft\helpers\DateTimeHelper;
-use craft\helpers\StringHelper;
-use GuzzleHttp\Client as HttpClient;
-use percipioglobal\notifications\channels\DatabaseChannel;
-use percipioglobal\notifications\channels\MailChannel;
-use percipioglobal\notifications\channels\SlackWebhookChannel;
-use percipioglobal\notifications\events\RegisterChannelsEvent;
-use percipioglobal\notifications\events\SendEvent;
 use Craft;
 use craft\base\Component;
-use percipioglobal\notifications\models\Notification;
-use percipioglobal\notifications\records\NotificationsRecord;
+use craft\elements\User;
+use craft\helpers\DateTimeHelper;
+use craft\helpers\Json;
+use craft\helpers\StringHelper;
+use GuzzleHttp\Client as HttpClient;
+use Illuminate\Support\Collection;
+use JetBrains\PhpStorm\ArrayShape;
+use JetBrains\PhpStorm\Pure;
+use percipiolondon\notifications\channels\DatabaseChannel;
+use percipiolondon\notifications\channels\MailChannel;
+use percipiolondon\notifications\channels\SlackWebhookChannel;
+use percipiolondon\notifications\events\RegisterChannelsEvent;
+use percipiolondon\notifications\events\SendEvent;
+use percipiolondon\notifications\models\Notification;
+use percipiolondon\notifications\records\NotificationsRecord;
 use yii\base\Event;
 use yii\base\InvalidCallException;
 
@@ -44,30 +48,30 @@ class NotificationsService extends Component
     /**
      * @event SubmissionEvent The event that is triggered before a notification is sent
      */
-    const EVENT_BEFORE_SEND = 'beforeSend';
+    public const EVENT_BEFORE_SEND = 'beforeSend';
 
     /**
      * @event SubmissionEvent The event that is triggered after a notification is sent
      */
-    const EVENT_AFTER_SEND = 'afterSend';
+    public const EVENT_AFTER_SEND = 'afterSend';
 
     /**
      * @event RegisterChannelsEvent An event to register new channels into the notificationsService
      */
-    const EVENT_REGISTER_CHANNELS = 'registerChannels';
+    public const EVENT_REGISTER_CHANNELS = 'registerChannels';
 
     // Public Methods
     // =========================================================================
     /**
      * @inheritdoc
      */
-    public static function channels()
+    public static function channels(): array
     {
         $channels = static::defineChannels();
 
         // Give plugins a chance to modify them
         $event = new RegisterChannelsEvent([
-            'channels' => $channels
+            'channels' => $channels,
         ]);
         Event::trigger(static::class, self::EVENT_REGISTER_CHANNELS, $event);
 
@@ -93,11 +97,11 @@ class NotificationsService extends Component
                 $notificationId = StringHelper::UUID();
                 $notification = clone $original;
 
-                if (! $notification->id) {
+                if (!$notification->id) {
                     $notification->id = $notificationId;
                 }
 
-                if (! $this->shouldSendNotification($notifiable, $notification, $channel)) {
+                if (!$this->shouldSendNotification($notifiable, $notification, $channel)) {
                     continue;
                 }
 
@@ -118,11 +122,11 @@ class NotificationsService extends Component
     /**
      * Get all notifications for a certain User
      *
-     * @param User $user
+     * @param User|null $user
      *
      * @return array
      */
-    public function getAll(User $user = null)
+    public function getAll(?User $user = null): array
     {
         // If there's no passed user, get the current logged in user
         $user = $user ?? Craft::$app->getUser();
@@ -139,11 +143,11 @@ class NotificationsService extends Component
     /**
      * Get all unread notifications for a certain User
      *
-     * @param User $user
+     * @param User|null $user
      *
      * @return array
      */
-    public function getAllUnread(User $user = null)
+    public function getAllUnread(User $user = null): array
     {
         // If there's no passed user, get the current logged in user
         $user = $user ?? Craft::$app->getUser();
@@ -156,6 +160,27 @@ class NotificationsService extends Component
         // No notifications when we don't have a passed in or logged in user
         return [];
     }
+
+	/**
+	 * Get all read notifications for a certain User
+	 *
+	 * @param User|null $user
+	 *
+	 * @return array
+	 */
+	public function getAllRead(User $user = null): array
+	{
+		// If there's no passed user, get the current logged in user
+		$user = $user ?? Craft::$app->getUser();
+
+		if ($user) {
+			$notifications = NotificationsRecord::find()->where(['notifiable' => $user->id])->andWhere(['not', ['read_at' => null]])->all();
+			return $this->formatNotificationData($notifications)->toArray();
+		}
+
+		// No notifications when we don't have a passed in or logged in user
+		return [];
+	}
 
     /**
      * Mark a notification as read
@@ -170,15 +195,19 @@ class NotificationsService extends Component
             $notifications = NotificationsRecord::find()->where(['notifiable' => $user->getId()])->all();
         }
 
-        // Make sure we have a collection to loop over
-        $notifications = collect($notifications);
+        if(is_array($notifications)) {
+                // Make sure we have a collection to loop over
+            $notifications = collect($notifications);
 
-        $notificationIds = $notifications->map(function ($notification) {
-            return is_object($notification) ? $notification->id : $notification;
-        });
+            $notificationIds = $notifications->map(function($notification) {
+                return is_object($notification) ? $notification->id : $notification;
+            });
+        } else {
+            $notificationIds = [$notifications->id];
+        }
 
         // Update the read notifications
-        if (! is_null($notificationIds)) {
+        if (!is_null($notificationIds)) {
             $now = DateTimeHelper::currentUTCDateTime()->format('Y-m-d H:i:s');
             NotificationsRecord::updateAll(['read_at' => $now], ['id' => $notificationIds]);
         }
@@ -189,12 +218,12 @@ class NotificationsService extends Component
      *
      * @param $notifications
      *
-     * @return \Illuminate\Support\Collection
+     * @return Collection
      */
-    protected function formatNotificationData($notifications)
+    protected function formatNotificationData($notifications): Collection
     {
-        return collect($notifications)->map(function ($notification) {
-            $notification->data = json_decode($notification->data);
+        return collect($notifications)->map(function($notification) {
+            $notification->data = Json::decode($notification->data);
             return $notification;
         });
     }
@@ -203,11 +232,11 @@ class NotificationsService extends Component
      * Determines if the notification can be sent.
      *
      * @param  mixed  $notifiable
-     * @param  Notification  $notification
-     * @param  string  $channel
+     * @param Notification $notification
+     * @param string $channel
      * @return bool
      */
-    protected function shouldSendNotification($notifiable, $notification, $channel)
+    protected function shouldSendNotification(mixed $notifiable, Notification $notification, string $channel): bool
     {
         $event = new SendEvent([
             'notification' => $notification,
@@ -226,7 +255,7 @@ class NotificationsService extends Component
      * @param  mixed  $notifiables
      * @return array
      */
-    protected function formatNotifiables($notifiables)
+    protected function formatNotifiables(mixed $notifiables): array
     {
         // Notifiables can be an anonymous function that returns an array
         if (is_callable($notifiables)) {
@@ -234,7 +263,7 @@ class NotificationsService extends Component
         }
 
         // Always make sure we have an array
-        if (! is_array($notifiables)) {
+        if (!is_array($notifiables)) {
             return [$notifiables];
         }
 
@@ -247,16 +276,16 @@ class NotificationsService extends Component
      * @return array The available channels.
      * @see channels()
      */
-    protected static function defineChannels()
+    protected static function defineChannels(): array
     {
         return [
-            'database' => function () {
+            'database' => function() {
                 return self::createDatabaseChannel();
             },
-            'mail' => function () {
+            'mail' => function() {
                 return self::createMailChannel();
             },
-            'slack' => function () {
+            'slack' => function() {
                 return self::createSlackChannel();
             },
         ];
@@ -265,10 +294,10 @@ class NotificationsService extends Component
     /**
      * Get a channel instance.
      *
-     * @param  string|null  $name
+     * @param string|null $name
      * @return mixed
      */
-    protected function channel($name = null)
+    protected function channel(?string $name = null): mixed
     {
         if (!isset(self::channels()[$name])) {
             throw new InvalidCallException("No channel {$name} exists.");
@@ -282,7 +311,7 @@ class NotificationsService extends Component
      *
      * @return DatabaseChannel
      */
-    protected static function createDatabaseChannel()
+    protected static function createDatabaseChannel(): DatabaseChannel
     {
         return new DatabaseChannel();
     }
@@ -292,7 +321,7 @@ class NotificationsService extends Component
      *
      * @return SlackWebhookChannel
      */
-    protected static function createSlackChannel()
+    protected static function createSlackChannel(): SlackWebhookChannel
     {
         return new SlackWebhookChannel(new HttpClient());
     }
@@ -302,7 +331,7 @@ class NotificationsService extends Component
      *
      * @return MailChannel
      */
-    protected static function createMailChannel()
+    protected static function createMailChannel(): MailChannel
     {
         return new MailChannel();
     }